requirements.txt
django-filter==24.3
Django==4.2.14
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9